+++ changelog.md
@@ -1,4 +1,4 @@
-# 26.1.2.1
+# 26.1.2.2
 ### Changes
-- Fixed info button in Uncrafting Table Screen and Auto Uncrafting Table Screen not rendering correctly
-- Restrictions list in common config no longer have default value
+- Added new config field `Restrict Ambiguously Crafted Items` to common config
+- Recipe that has ItemTags ingredient can be disabled by setting `Restrict Ambiguously Crafted Items` to `false`